requirements.txt
tweepy>=4.14.0
pillow>=10.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.1